use in-process grpc client instead of wrapping server interface
