+++ go.mod
@@ -484,8 +484,12 @@ require k8s.io/code-generator v0.29.1 // @grafana/grafana-app-platform-squad
 
 require github.com/spyzhov/ajson v0.9.0 // @grafana/grafana-app-platform-squad
 
+require github.com/fullstorydev/grpchan v1.1.1 // @grafana/backend-platform
+
 require (
+	github.com/bufbuild/protocompile v0.4.0 // indirect
 	github.com/grafana/sqlds/v3 v3.2.0 // indirect
+	github.com/jhump/protoreflect v1.15.1 // indirect
 	github.com/mitchellh/go-homedir v1.1.0 // indirect
 	github.com/mithrandie/csvq v1.17.10 // indirect
 	github.com/mithrandie/csvq-driver v1.6.8 // indirect
